# port fxhash from rust to zig