Processing time decreased about 6ms. (from 28ms to 22ms in my computer)