@jinyus 2023-10-27
Main:
`Zig | 28.38 ms | 430.33 ms | 3.79 s`

This PR:
```
Zig:

        Benchmark 1: ./main
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
        Processing time (w/o IO): 23ms
          Time (mean ± σ):      70.3 ms ±   0.5 ms    [User: 64.9 ms, System: 5.3 ms]
          Range (min … max):    69.7 ms …  71.5 ms    10 runs
         
Zig:

        Benchmark 1: ./main
        Processing time (w/o IO): 340ms
        Processing time (w/o IO): 340ms
        Processing time (w/o IO): 339ms
          Time (mean ± σ):     530.1 ms ±   0.3 ms    [User: 500.6 ms, System: 29.5 ms]
          Range (min … max):   529.9 ms … 530.4 ms    2 runs
         
Zig:

        Benchmark 1: ./main
        Processing time (w/o IO): 3017ms
        Processing time (w/o IO): 3013ms
        Processing time (w/o IO): 3026ms
          Time (mean ± σ):      3.620 s ±  0.001 s    [User: 3.551 s, System: 0.062 s]
          Range (min … max):    3.619 s …  3.621 s    2 runs
         

```